Rewrite of the Linux load meter.
Prior to the rewrite the displayed scale range was unpredictable making it
impossible to estimate the system load average when usedLabels was not
enabled.
  The scale range was determined as follows:
    + The scale range is always >= 1.0
    + If the current load value is > the current scale range,
        the scale range becomes 5 times the current load value.
    + If the current load value < 1/5 the current scale range,
        the scale range becomes the greater of 1.0 and the current load
        value.
  The bar color was determined as follows with the default resources:
    + The color of the unused part of the display was set to "loadIdleColor".
    + If the current load average was less than or equal to the CPU count,
        the color was set to loadProcColor.
    + If the current load average was less than or equal to 4 * CPU count,
        the color was set to "loadWarnColor".
    + Otherwise, the the color was set to "loadCritColor".

After the rewrite the displayed scale range is predictable and the bar
also indicates what the displayed scale range is.

Theory:

The current load avgerage value (L) is converted into a form like
exponential or scientific noatation. The components of this form are:
  + Fract:  "fractional part"; real; [0, 1]
  + B:      "base for the exponential multiplier"; real; (1, infinity)
  + E:      "power for the exponential multiplier"; integer; [0, infinity)
Such that:
  + L = Fract * B^E
  + Fract is the largest value that conforms to the above constraints.
If E is greater than or equal to the number of colors listed in the
"loadColorList", the meter is "pegged"; E is set to the color count - 1
and Fract is set to 1.0.

The high end of the displayed scale range is B^E and E determines the bar
color.

With the default Resources, the meanings of the colors has not changed;
only how the scale range is determined.

Behaviors:

  The following display behaviors are avaialabe and selected by the
  "loadScalingBehavior" resource:
    + "Traditional"
        This behavior is similar to what the old code did.
          + The low end of the displayed scale range is 0.0.
          + The bar length is set to Fract.
          + The color of the "idle" part of the display is set to
              "loadIdleColor".
    + "Overlay"
        This behavior provides a display that appears to grow from the low
            end of the scale, as the current load avgerage increases,
            overlaying the previous display.
          + If E is 0:
            + The low end of the displayed scale range is 0.
            + The bar length is Fract.
            + The color of the "idle" part of the display is set to
                "loadIdleColor".
          + IF is not 0:
            + The low end of the displayed scale range is E/B.
            + The bar length is Fract - 1/B.
            + The color of the "idle" part of the display is set to
                the bar color for E-1.

Since the only Linux specific part of the code is how the current system
load average is obtained, porting to other platforms should be easy.